requirements.txt
opencv-python
torch
numpy
playsound
piper-tts
wave
pydub
simpleaudio
ultralytics